Organize the serializer related code better
