+++ lib/acidic_job.rb
@@ -3,6 +3,7 @@
 require_relative "acidic_job/version"
 require_relative "acidic_job/errors"
 require_relative "acidic_job/logger"
+require_relative "acidic_job/arguments"
 require_relative "acidic_job/serializer"
 require_relative "acidic_job/workflow_builder"
 require_relative "acidic_job/idempotency_key"
@@ -15,5 +16,21 @@ require_relative "acidic_job/processor"
 require_relative "acidic_job/mixin"
 require_relative "acidic_job/base"
 
+require_relative "acidic_job/serializers/exception_serializer"
+require_relative "acidic_job/serializers/finished_point_serializer"
+require_relative "acidic_job/serializers/job_serializer"
+require_relative "acidic_job/serializers/range_serializer"
+require_relative "acidic_job/serializers/recovery_point_serializer"
+require_relative "acidic_job/serializers/worker_serializer"
+require "active_job/serializers"
+
 module AcidicJob
+  ::ActiveJob::Serializers.add_serializers(
+    Serializers::ExceptionSerializer,
+    Serializers::FinishedPointSerializer,
+    Serializers::JobSerializer,
+    Serializers::RangeSerializer,
+    Serializers::RecoveryPointSerializer,
+    Serializers::WorkerSerializer
+  )
 end
